requirements.txt
# feedparser is only if you want to scrape arxiv
feedparser
numpy
scipy

# sciki-learn is only for sparse arrays
scikit-learn

# twitter
python-twitter
pytz

# the following are only for serve.py
python-dateutil
flask
flask_limiter
tornado
pymongo
mongoengine
python-dotenv
bs4
black
dnspython
meilisearch
